fix: dash-generate-components wont scrape imported files for props
