+++ src/lib/components/ExtendableGraph.react.js
@@ -1,22 +1,11 @@
 import React, {Component, Suspense, memo} from 'react';
 import PropTypes from 'prop-types';
 
-import {mergeDeepRight, isNil, type, omit, equals} from 'ramda';
-import ResizeDetector from 'react-resize-detector';
 import {
-    graphPropTypes,
-    graphDefaultProps,
+    privatePropTypes,
+    privateDefaultProps,
 } from '../fragments/ExtendableGraph.props';
-import {
-    AUTO_LAYOUT,
-    RESPONSIVE_LAYOUT,
-    UNRESPONSIVE_LAYOUT,
-    AUTO_CONFIG,
-    RESPONSIVE_CONFIG,
-    UNRESPONSIVE_CONFIG,
-} from '../fragments/ExtendableGraph.responsiveprops';
-import {filterEventData} from '../fragments/ExtendableGraph.events';
-/* global Plotly:true */
+import graph from '../fragments/ExtendableGraph.react';
 
 const EMPTY_DATA = [];
 
@@ -123,6 +112,8 @@ class GraphWithCustomProps extends Component {
     }
 }
 
+const Graph = graph;
+
 const GraphContainer = memo((props) => {
     const {className, id} = props;
 
@@ -145,432 +136,430 @@ const GraphContainer = memo((props) => {
     );
 });
 
-/**
- * ExtendableGraph can be used to render any plotly.js-powered data vis.
- */
-class Graph extends Component {
-    constructor(props) {
-        super(props);
-        this.gd = React.createRef();
-        this._hasPlotted = false;
-        this._prevGd = null;
+export const graphPropTypes = {
+    ...privatePropTypes,
 
-        this.bindEvents = this.bindEvents.bind(this);
-        this.getConfig = this.getConfig.bind(this);
-        this.getConfigOverride = this.getConfigOverride.bind(this);
-        this.getLayout = this.getLayout.bind(this);
-        this.getLayoutOverride = this.getLayoutOverride.bind(this);
-        this.graphResize = this.graphResize.bind(this);
-        this.isResponsive = this.isResponsive.bind(this);
+    /**
+     * The ID of this component, used to identify dash components
+     * in callbacks. The ID needs to be unique across all of the
+     * components in an app.
+     */
+    id: PropTypes.string,
 
-        this.state = {override: {}, originals: {}};
-    }
+    /**
+     * If True, the Plotly.js plot will be fully responsive to window resize
+     * and parent element resize event. This is achieved by overriding
+     * `config.responsive` to True, `figure.layout.autosize` to True and unsetting
+     * `figure.layout.height` and `figure.layout.width`.
+     * If False, the Plotly.js plot not be responsive to window resize and
+     * parent element resize event. This is achieved by overriding `config.responsive`
+     * to False and `figure.layout.autosize` to False.
+     * If 'auto' (default), the Graph will determine if the Plotly.js plot can be made fully
+     * responsive (True) or not (False) based on the values in `config.responsive`,
+     * `figure.layout.autosize`, `figure.layout.height`, `figure.layout.width`.
+     * This is the legacy behavior of the ExtendableGraph component.
+     *
+     * Needs to be combined with appropriate dimension / styling through the `style` prop
+     * to fully take effect.
+     */
+    responsive: PropTypes.oneOf([true, false, 'auto']),
 
-    plot(props) {
-        let {figure, config} = props;
-        const {animate, animation_options, responsive} = props;
-        const gd = this.gd.current;
+    /**
+     * Data from latest click event. Read-only.
+     */
+    clickData: PropTypes.object,
 
-        figure = props._dashprivate_transformFigure(figure, gd);
-        config = props._dashprivate_transformConfig(config, gd);
+    /**
+     * Data from latest click annotation event. Read-only.
+     */
+    clickAnnotationData: PropTypes.object,
 
-        if (
-            animate &&
-            this._hasPlotted &&
-            figure.data.length === gd.data.length
-        ) {
-            return Plotly.animate(gd, figure, animation_options);
-        }
+    /**
+     * Data from latest hover event. Read-only.
+     */
+    hoverData: PropTypes.object,
 
-        const configClone = this.getConfig(config, responsive);
-        const layoutClone = this.getLayout(figure.layout, responsive);
+    /**
+     * If True, `clear_on_unhover` will clear the `hoverData` property
+     * when the user "unhovers" from a point.
+     * If False, then the `hoverData` property will be equal to the
+     * data from the last point that was hovered over.
+     */
+    clear_on_unhover: PropTypes.bool,
 
-        gd.classList.add('dash-graph--pending');
+    /**
+     * Data from latest select event. Read-only.
+     */
+    selectedData: PropTypes.object,
 
-        return Plotly.react(gd, {
-            data: figure.data,
-            layout: layoutClone,
-            frames: figure.frames,
-            config: configClone,
-        }).then(() => {
-            const gd = this.gd.current;
+    /**
+     * Data from latest relayout event which occurs
+     * when the user zooms or pans on the plot or other
+     * layout-level edits. Has the form `{<attr string>: <value>}`
+     * describing the changes made. Read-only.
+     */
+    relayoutData: PropTypes.object,
 
-            // double-check gd hasn't been unmounted
-            if (!gd) {
-                return;
-            }
+    /**
+     * Data that should be appended to existing traces. Has the form
+     * `[updateData, traceIndices, maxPoints]`, where `updateData` is an array
+     * containing data objects to extend, `traceIndices` (optional) is an array
+     * of trace indices that should be extended, and `maxPoints` (optional) is
+     * either an integer defining the maximum number of points allowed or an
+     * object with key:value pairs matching `updateData`
+     * Reference the Plotly.extendTraces API for full usage:
+     * https://plot.ly/javascript/plotlyjs-function-reference/#plotlyextendtraces
+     */
+    extendData: PropTypes.oneOfType([PropTypes.array, PropTypes.object]),
 
-            gd.classList.remove('dash-graph--pending');
+    /**
+     * Data that should be prepended to existing traces. Has the form
+     * `[updateData, traceIndices, maxPoints]`, where `updateData` is an array
+     * containing data objects to prepend, `traceIndices` (optional) is an array
+     * of trace indices that should be extended, and `maxPoints` (optional) is
+     * either an integer defining the maximum number of points allowed or an
+     * object with key:value pairs matching `updateData`
+     * Reference the Plotly.prependTraces API for full usage:
+     * https://plot.ly/javascript/plotlyjs-function-reference/#plotlyprependtraces
+     */
+    prependData: PropTypes.oneOfType([PropTypes.array, PropTypes.object]),
 
-            // in case we've made a new DOM element, transfer events
-            if (this._hasPlotted && gd !== this._prevGd) {
-                if (this._prevGd && this._prevGd.removeAllListeners) {
-                    this._prevGd.removeAllListeners();
-                    Plotly.purge(this._prevGd);
-                }
-                this._hasPlotted = false;
-            }
+    /**
+     * Data from latest restyle event which occurs
+     * when the user toggles a legend item, changes
+     * parcoords selections, or other trace-level edits.
+     * Has the form `[edits, indices]`, where `edits` is an object
+     * `{<attr string>: <value>}` describing the changes made,
+     * and `indices` is an array of trace indices that were edited.
+     * Read-only.
+     */
+    restyleData: PropTypes.array,
 
-            if (!this._hasPlotted) {
-                this.bindEvents();
-                this.graphResize(true);
-                this._hasPlotted = true;
-                this._prevGd = gd;
-            }
-        });
-    }
+    /**
+     * Plotly `figure` object. See schema:
+     * https://plot.ly/javascript/reference
+     *
+     * `config` is set separately by the `config` property
+     */
+    figure: PropTypes.exact({
+        data: PropTypes.arrayOf(PropTypes.object),
+        layout: PropTypes.object,
+        frames: PropTypes.arrayOf(PropTypes.object),
+    }),
 
-    mergeTraces(props, dataKey, plotlyFnKey) {
-        const data = props[dataKey];
-        const gd = this.gd.current;
+    /**
+     * Generic style overrides on the plot div
+     */
+    style: PropTypes.object,
 
-        let updateData, traceIndices, maxPoints;
+    /**
+     * className of the parent div
+     */
+    className: PropTypes.string,
 
-        if (gd.data.length < 1) {
-            // figure has no pre-existing data. redirect to plot()
-            props.figure.data = data[0];
-            props.clearState(dataKey);
-            this.plot(props);
-            return;
-        }
+    /**
+     * Beta: If true, animate between updates using
+     * plotly.js's `animate` function
+     */
+    animate: PropTypes.bool,
 
-        if (Array.isArray(data) && Array.isArray(data[0])) {
-            [updateData, traceIndices, maxPoints] = data;
-        } else {
-            updateData = data;
-        }
+    /**
+     * Beta: Object containing animation settings.
+     * Only applies if `animate` is `true`
+     */
+    animation_options: PropTypes.object,
 
-        // if no traceIndices specified, generate them based on the length of the input data
-        if (!traceIndices) {
-            traceIndices = Array.from(Array(updateData.length).keys());
-        }
+    /**
+     * Plotly.js config options.
+     * See https://plot.ly/javascript/configuration-options/
+     * for more info.
+     */
+    config: PropTypes.exact({
+        /**
+         * No interactivity, for export or image generation
+         */
+        staticPlot: PropTypes.bool,
 
-        function createDataObject(data) {
-            const dataprops = Object.keys(data);
-            const ret = {};
-            for (let i = 0; i < dataprops.length; i++) {
-                ret[dataprops[i]] = [data[dataprops[i]]];
-            }
-            return ret;
-        }
+        /**
+         * Base URL for a Plotly cloud instance, if `showSendToCloud` is enabled
+         */
+        plotlyServerURL: PropTypes.string,
 
-        for (const [i, value] of updateData.entries()) {
-            // extend/add each trace within updateData
-            const updateObject = createDataObject(value);
-            if (i < updateData.length - 1) {
-                if (traceIndices[i] < gd.data.length) {
-                    Plotly[plotlyFnKey](
-                        gd,
-                        updateObject,
-                        [traceIndices[i]],
-                        maxPoints
-                    );
-                } else {
-                    Plotly.addTraces(gd, value);
-                }
-            } else {
-                if (traceIndices[i] < gd.data.length) {
-                    Plotly[plotlyFnKey](
-                        gd,
-                        updateObject,
-                        [traceIndices[i]],
-                        maxPoints
-                    ).then(() => {
-                        const gd = this.gd.current;
+        /**
+         * We can edit titles, move annotations, etc - sets all pieces of `edits`
+         * unless a separate `edits` config item overrides individual parts
+         */
+        editable: PropTypes.bool,
 
-                        // double-check gd hasn't been unmounted
-                        if (!gd) {
-                            return;
-                        }
-                        gd.classList.remove('dash-graph--pending');
-                    });
-                } else {
-                    Plotly.addTraces(gd, value).then(() => {
-                        const gd = this.gd.current;
-
-                        // double-check gd hasn't been unmounted
-                        if (!gd) {
-                            return;
-                        }
-                        gd.classList.remove('dash-graph--pending');
-                    });
-                }
-            }
-        }
-        props.clearState(dataKey);
-        return;
-    }
-
-    getConfig(config, responsive) {
-        return mergeDeepRight(config, this.getConfigOverride(responsive));
-    }
-
-    getLayout(layout, responsive) {
-        if (!layout) {
-            return layout;
-        }
-        const override = this.getLayoutOverride(responsive);
-        const {override: prev_override, originals: prev_originals} = this.state;
-        // Store the original data that we're about to override
-        const originals = {};
-        for (const key in override) {
-            if (layout[key] !== prev_override[key]) {
-                originals[key] = layout[key];
-            } else if (
-                Object.prototype.hasOwnProperty.call(prev_originals, key)
-            ) {
-                originals[key] = prev_originals[key];
-            }
-        }
-        this.setState({override, originals});
-        // Undo the previous override, but only for keys that the user did not change
-        for (const key in prev_originals) {
-            if (layout[key] === prev_override[key]) {
-                layout[key] = prev_originals[key];
-            }
-        }
-        // Apply the current override
-        for (const key in override) {
-            layout[key] = override[key];
-        }
-        return layout; // not really a clone
-    }
-
-    getConfigOverride(responsive) {
-        switch (responsive) {
-            case false:
-                return UNRESPONSIVE_CONFIG;
-            case true:
-                return RESPONSIVE_CONFIG;
-            default:
-                return AUTO_CONFIG;
-        }
-    }
-
-    getLayoutOverride(responsive) {
-        switch (responsive) {
-            case false:
-                return UNRESPONSIVE_LAYOUT;
-            case true:
-                return RESPONSIVE_LAYOUT;
-            default:
-                return AUTO_LAYOUT;
-        }
-    }
-
-    isResponsive(props) {
-        const {config, figure, responsive} = props;
-
-        if (type(responsive) === 'Boolean') {
-            return responsive;
-        }
-
-        return Boolean(
-            config.responsive &&
-                (!figure.layout ||
-                    ((figure.layout.autosize ||
-                        isNil(figure.layout.autosize)) &&
-                        (isNil(figure.layout.height) ||
-                            isNil(figure.layout.width))))
-        );
-    }
-
-    graphResize(force = false) {
-        if (!force && !this.isResponsive(this.props)) {
-            return;
-        }
-
-        const gd = this.gd.current;
-        if (!gd) {
-            return;
-        }
-
-        gd.classList.add('dash-graph--pending');
-        Plotly.Plots.resize(gd)
-            .catch(() => {})
-            .finally(() => gd.classList.remove('dash-graph--pending'));
-    }
-
-    bindEvents() {
-        const {
-            setProps,
-            clear_on_unhover,
-            relayoutData,
-            restyleData,
-            hoverData,
-            selectedData,
-        } = this.props;
-
-        const gd = this.gd.current;
-
-        gd.on('plotly_click', (eventData) => {
-            const clickData = filterEventData(gd, eventData, 'click');
-            if (!isNil(clickData)) {
-                setProps({clickData});
-            }
-        });
-        gd.on('plotly_clickannotation', (eventData) => {
-            const clickAnnotationData = omit(
-                ['event', 'fullAnnotation'],
-                eventData
-            );
-            setProps({clickAnnotationData});
-        });
-        gd.on('plotly_hover', (eventData) => {
-            const hover = filterEventData(gd, eventData, 'hover');
-            if (!isNil(hover) && !equals(hover, hoverData)) {
-                setProps({hoverData: hover});
-            }
-        });
-        gd.on('plotly_selected', (eventData) => {
-            const selected = filterEventData(gd, eventData, 'selected');
-            if (!isNil(selected) && !equals(selected, selectedData)) {
-                setProps({selectedData: selected});
-            }
-        });
-        gd.on('plotly_deselect', () => {
-            setProps({selectedData: null});
-        });
-        gd.on('plotly_relayout', (eventData) => {
-            const relayout = filterEventData(gd, eventData, 'relayout');
-            if (!isNil(relayout) && !equals(relayout, relayoutData)) {
-                setProps({relayoutData: relayout});
-            }
-        });
-        gd.on('plotly_restyle', (eventData) => {
-            const restyle = filterEventData(gd, eventData, 'restyle');
-            if (!isNil(restyle) && !equals(restyle, restyleData)) {
-                setProps({restyleData: restyle});
-            }
-        });
-        gd.on('plotly_unhover', () => {
-            if (clear_on_unhover) {
-                setProps({hoverData: null});
-            }
-        });
-    }
-
-    componentDidMount() {
-        this.plot(this.props);
-        if (this.props.prependData) {
-            this.mergeTraces(this.props, 'prependData', 'prependTraces');
-        }
-        if (this.props.extendData) {
-            this.mergeTraces(this.props, 'extendData', 'extendTraces');
-        }
-        if (this.props.prependData?.length || this.props.extendData?.length) {
-            this.props._dashprivate_onFigureModified(this.props.figure);
-        }
-    }
-
-    componentWillUnmount() {
-        const gd = this.gd.current;
-        if (gd && gd.removeAllListeners) {
-            gd.removeAllListeners();
-            if (this._hasPlotted) {
-                Plotly.purge(gd);
-            }
-        }
-    }
-
-    shouldComponentUpdate(nextProps) {
-        return (
-            this.props.id !== nextProps.id ||
-            JSON.stringify(this.props.style) !==
-                JSON.stringify(nextProps.style) ||
-            JSON.stringify(this.props.loading_state) !==
-                JSON.stringify(nextProps.loading_state)
-        );
-    }
-
-    UNSAFE_componentWillReceiveProps(nextProps) {
-        const idChanged = this.props.id !== nextProps.id;
-        if (idChanged) {
-            /*
-             * then the dom needs to get re-rendered with a new ID.
-             * the graph will get updated in componentDidUpdate
+        /**
+         * A set of editable properties
+         */
+        edits: PropTypes.exact({
+            /**
+             * The main anchor of the annotation, which is the
+             * text (if no arrow) or the arrow (which drags the whole thing leaving
+             * the arrow length & direction unchanged)
              */
-            return;
-        }
-        if (
-            this.props.figure !== nextProps.figure ||
-            this.props._dashprivate_transformConfig !==
-                nextProps._dashprivate_transformConfig ||
-            this.props._dashprivate_transformFigure !==
-                nextProps._dashprivate_transformFigure
-        ) {
-            this.plot(nextProps);
-        }
+            annotationPosition: PropTypes.bool,
 
-        if (this.props.prependData !== nextProps.prependData) {
-            this.mergeTraces(nextProps, 'prependData', 'prependTraces');
-        }
+            /**
+             * Just for annotations with arrows, change the length and direction of the arrow
+             */
+            annotationTail: PropTypes.bool,
 
-        if (this.props.extendData !== nextProps.extendData) {
-            this.mergeTraces(nextProps, 'extendData', 'extendTraces');
-        }
+            annotationText: PropTypes.bool,
 
-        if (this.props.prependData?.length || this.props.extendData?.length) {
-            this.props._dashprivate_onFigureModified(this.props.figure);
-        }
-    }
+            axisTitleText: PropTypes.bool,
 
-    componentDidUpdate(prevProps) {
-        if (prevProps.id !== this.props.id) {
-            this.plot(this.props);
-        }
-    }
+            colorbarPosition: PropTypes.bool,
 
-    render() {
-        const {className, id, style, loading_state} = this.props;
+            colorbarTitleText: PropTypes.bool,
 
-        return (
-            <div
-                data-dash-is-loading={
-                    (loading_state && loading_state.is_loading) || undefined
-                }
-                style={style}
-                className={className}
-            >
-                <ResizeDetector
-                    handleHeight={true}
-                    handleWidth={true}
-                    refreshMode="debounce"
-                    refreshOptions={{trailing: true}}
-                    refreshRate={50}
-                    onResize={this.graphResize}
-                />
-                <div
-                    id={id}
-                    key={id}
-                    ref={this.gd}
-                    style={{height: '100%', width: '100%'}}
-                />
-            </div>
-        );
-    }
-}
-Graph.propTypes = {
-    ...graphPropTypes,
-    prependData: PropTypes.arrayOf(
-        PropTypes.oneOfType([PropTypes.array, PropTypes.object])
-    ),
-    extendData: PropTypes.arrayOf(
-        PropTypes.oneOfType([PropTypes.array, PropTypes.object])
-    ),
-    clearState: PropTypes.func.isRequired,
+            legendPosition: PropTypes.bool,
+
+            /**
+             * Edit the trace name fields from the legend
+             */
+            legendText: PropTypes.bool,
+
+            shapePosition: PropTypes.bool,
+
+            /**
+             * The global `layout.title`
+             */
+            titleText: PropTypes.bool,
+        }),
+
+        /**
+         * DO autosize once regardless of layout.autosize
+         * (use default width or height values otherwise)
+         */
+        autosizable: PropTypes.bool,
+
+        /**
+         * Whether to change layout size when the window size changes
+         */
+        responsive: PropTypes.bool,
+
+        /**
+         * Set the length of the undo/redo queue
+         */
+        queueLength: PropTypes.number,
+
+        /**
+         * If we DO autosize, do we fill the container or the screen?
+         */
+        fillFrame: PropTypes.bool,
+
+        /**
+         * If we DO autosize, set the frame margins in percents of plot size
+         */
+        frameMargins: PropTypes.number,
+
+        /**
+         * Mousewheel or two-finger scroll zooms the plot
+         */
+        scrollZoom: PropTypes.bool,
+
+        /**
+         * Double click interaction (false, 'reset', 'autosize' or 'reset+autosize')
+         */
+        doubleClick: PropTypes.oneOf([
+            false,
+            'reset',
+            'autosize',
+            'reset+autosize',
+        ]),
+
+        /**
+         * New users see some hints about interactivity
+         */
+        showTips: PropTypes.bool,
+
+        /**
+         * Enable axis pan/zoom drag handles
+         */
+        showAxisDragHandles: PropTypes.bool,
+
+        /**
+         * Enable direct range entry at the pan/zoom drag points
+         * (drag handles must be enabled above)
+         */
+        showAxisRangeEntryBoxes: PropTypes.bool,
+
+        /**
+         * Link to open this plot in plotly
+         */
+        showLink: PropTypes.bool,
+
+        /**
+         * If we show a link, does it contain data or just link to a plotly file?
+         */
+        sendData: PropTypes.bool,
+
+        /**
+         * Text appearing in the sendData link
+         */
+        linkText: PropTypes.string,
+
+        /**
+         * Display the mode bar (true, false, or 'hover')
+         */
+        displayModeBar: PropTypes.oneOf([true, false, 'hover']),
+
+        /**
+         * Should we include a modebar button to send this data to a
+         * Plotly Cloud instance, linked by `plotlyServerURL`.
+         * By default this is false.
+         */
+        showSendToCloud: PropTypes.bool,
+
+        /**
+         * Remove mode bar button by name.
+         * All modebar button names at https://github.com/plotly/plotly.js/blob/master/src/components/modebar/buttons.js
+         * Common names include:
+         * sendDataToCloud;
+         * (2D) zoom2d, pan2d, select2d, lasso2d, zoomIn2d, zoomOut2d, autoScale2d, resetScale2d;
+         * (Cartesian) hoverClosestCartesian, hoverCompareCartesian;
+         * (3D) zoom3d, pan3d, orbitRotation, tableRotation, handleDrag3d, resetCameraDefault3d, resetCameraLastSave3d, hoverClosest3d;
+         * (Geo) zoomInGeo, zoomOutGeo, resetGeo, hoverClosestGeo;
+         * hoverClosestGl2d, hoverClosestPie, toggleHover, resetViews.
+         */
+        modeBarButtonsToRemove: PropTypes.array,
+
+        /**
+         * Add mode bar button using config objects
+         */
+        modeBarButtonsToAdd: PropTypes.array,
+
+        /**
+         * Fully custom mode bar buttons as nested array,
+         * where the outer arrays represents button groups, and
+         * the inner arrays have buttons config objects or names of default buttons
+         */
+        modeBarButtons: PropTypes.any,
+
+        /**
+         * Modifications to how the toImage modebar button works
+         */
+        toImageButtonOptions: PropTypes.exact({
+            /**
+             * The file format to create
+             */
+            format: PropTypes.oneOf(['jpeg', 'png', 'webp', 'svg']),
+            /**
+             * The name given to the downloaded file
+             */
+            filename: PropTypes.string,
+            /**
+             * Width of the downloaded file, in px
+             */
+            width: PropTypes.number,
+            /**
+             * Height of the downloaded file, in px
+             */
+            height: PropTypes.number,
+            /**
+             * Extra resolution to give the file after
+             * rendering it with the given width and height
+             */
+            scale: PropTypes.number,
+        }),
+
+        /**
+         * Add the plotly logo on the end of the mode bar
+         */
+        displaylogo: PropTypes.bool,
+
+        /**
+         * Add the plotly logo even with no modebar
+         */
+        watermark: PropTypes.bool,
+
+        /**
+         * Increase the pixel ratio for Gl plot images
+         */
+        plotGlPixelRatio: PropTypes.number,
+
+        /**
+         * URL to topojson files used in geo charts
+         */
+        topojsonURL: PropTypes.string,
+
+        /**
+         * Mapbox access token (required to plot mapbox trace types)
+         * If using an Mapbox Atlas server, set this option to '',
+         * so that plotly.js won't attempt to authenticate to the public Mapbox server.
+         */
+        mapboxAccessToken: PropTypes.any,
+
+        /**
+         * The locale to use. Locales may be provided with the plot
+         * (`locales` below) or by loading them on the page, see:
+         * https://github.com/plotly/plotly.js/blob/master/dist/README.md#to-include-localization
+         */
+        locale: PropTypes.string,
+
+        /**
+         * Localization definitions, if you choose to provide them with the
+         * plot rather than registering them globally.
+         */
+        locales: PropTypes.object,
+    }),
+
+    /**
+     * Function that updates the state tree.
+     */
+    setProps: PropTypes.func,
+
+    /**
+     * Object that holds the loading state object coming from dash-renderer
+     */
+    loading_state: PropTypes.shape({
+        /**
+         * Determines if the component is loading or not
+         */
+        is_loading: PropTypes.bool,
+        /**
+         * Holds which property is loading
+         */
+        prop_name: PropTypes.string,
+        /**
+         * Holds the name of the component that is loading
+         */
+        component_name: PropTypes.string,
+    }),
 };
-Graph.defaultProps = {
-    ...graphDefaultProps,
-    prependData: [],
-    extendData: [],
+
+export const graphDefaultProps = {
+    ...privateDefaultProps,
+    clickData: null,
+    clickAnnotationData: null,
+    hoverData: null,
+    selectedData: null,
+    relayoutData: null,
+    extendData: null,
+    prependData: null,
+    restyleData: null,
+    figure: {data: [], layout: {}, frames: []},
+    responsive: 'auto',
+    animate: false,
+    animation_options: {
+        frame: {
+            redraw: false,
+        },
+        transition: {
+            duration: 750,
+            ease: 'cubic-in-out',
+        },
+    },
+    clear_on_unhover: false,
+    config: {},
 };
 
 GraphWithCustomProps.propTypes = graphPropTypes;
 GraphContainer.propTypes = graphPropTypes;
 
 GraphWithCustomProps.defaultProps = graphDefaultProps;
-
-
 
 export default GraphWithCustomProps;
